+++ src/test/scala/com/jcdecaux/setl/MockCassandra.scala
@@ -2,10 +2,18 @@ package com.jcdecaux.setl
 
 import com.datastax.driver.core.Session
 import com.datastax.spark.connector.cql.CassandraConnector
+import org.apache.spark.SparkConf
 
 class MockCassandra(connector: CassandraConnector, keyspace: String) {
 
   private def dropTable(table: String, session: Session): Unit = session.execute(s"DROP TABLE IF EXISTS $keyspace.$table;")
+
+  def dropKeyspace(): this.type = {
+    connector.withSessionDo(session => {
+      session.execute(s"DROP KEYSPACE IF EXISTS $keyspace;")
+    })
+    this
+  }
 
   def generateKeyspace(): this.type = {
     connector.withSessionDo(session => {
@@ -313,4 +321,16 @@ class MockCassandra(connector: CassandraConnector, keyspace: String) {
 
 object MockCassandra {
   val keyspace: String = "test_space"
+  val host: String = System.getProperty("setl.test.cassandra.host", "localhost")
+
+  val cassandraConf: SparkConf = new SparkConf(true)
+    .set("spark.cassandra.connection.host", host)
+    .set("spark.cassandra.connection.port", "9042")
+    .set("spark.cassandra.connection.keep_alive_ms", "5000")
+    .set("spark.cassandra.connection.timeout_ms", "30000")
+    .set("spark.ui.showConsoleProgress", "false")
+    .set("spark.ui.enabled", "false")
+    .set("spark.cleaner.ttl", "3600")
+    .setMaster("local[*]")
+    .setAppName("Test")
 }